split options out from main publisher/listener files
